Remove esphome Param stuff
